Move `SidebarNotifications` into the `AppSidebar` component
